requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14,<4.0
psycopg2-binary>=2.9,<3.0
python-decouple>=3.8
drf-spectacular>=0.26,<1.0
django-cors-headers>=4.0,<5.0
django-ratelimit>=4.1,<5.0
gunicorn>=21.0,<22.0
